+++ src/main/java/org/jamesframework/examples/tsp2/TSP2.java
@@ -18,6 +18,8 @@ package org.jamesframework.examples.tsp2;
 
 import java.io.FileNotFoundException;
 import java.text.DecimalFormat;
+import java.text.DecimalFormatSymbols;
+import java.util.Locale;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import org.jamesframework.core.problems.objectives.evaluations.Evaluation;
@@ -174,7 +176,7 @@ public class TSP2 {
             System.out.println("Time limit: " + timeLimit + " seconds");
             System.out.println("---------------------------------------");
 
-            DecimalFormat df = new DecimalFormat("0.0");
+            DecimalFormat df = new DecimalFormat("0.0", DecimalFormatSymbols.getInstance(Locale.US));
             System.out.format("%20s    %15s \n", "", "Travel distance");
             System.out.format("%20s    %15s \n",
                                 "Random descent:",